Feature/item position fixes (#2)
* - Resolved item position issue
- Updated library version
- Changed version

* - Managed manual scroll scenario
