+++ app/src/main/java/com/mindinventory/autoscrollviewpager/MainActivity.kt
@@ -1,6 +1,8 @@
 package com.mindinventory.autoscrollviewpager
 
+import android.os.Build
 import android.os.Bundle
+import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
 
@@ -14,6 +16,7 @@ class MainActivity : AppCompatActivity() {
     }
     private val myAdapter by lazy { MyAdapter() }
 
+    @RequiresApi(Build.VERSION_CODES.Q)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
